pyproject.toml
[project]
name = "knowledge-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==25.1.0",
    "aiohttp>=3.13.3",
    "aiosqlite==0.22.1",
    "beautifulsoup4>=4.14.3",
    "bilibili-api-python>=17.4.1",
    "chromadb>=1.5.5",
    "click==8.3.1",
    "databases>=0.9.0",
    "docx2txt>=0.9",
    "email-validator>=2.3.0",
    "emoji==2.15.0",
    "fastapi==0.135.1",
    "fastmcp>=3.1.0",
    "filetype>=1.2.0",
    "httpx[http2,socks]==0.28.1",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "lxml==6.0.2",
    "openai-whisper>=20250625",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "pyperclip==1.11.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "pyyaml==6.0.3",
    "redis>=7.4.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
    "sqlalchemy-utils>=0.42.1",
    "srt>=3.5.3",
    "starlette>=1.0.0",
    "tenacity>=9.1.4",
    "textual==8.1.1",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "uvicorn==0.42.0",
    "websockets==16.0",
]
